Move utils, io, and stats back up a level.
Having all three submodules in another submodule made importing
problematic, because nimare.dataset needs nimare.utils.utils but
nimare.utils.io needs nimare.dataset.
