+++ nimare/__init__.py
@@ -8,4 +8,7 @@ with warnings.catch_warnings(record=True) as w:
     from . import dataset
     from . import meta
     from . import resources
+    from . import io
+    from . import stats
+    from . import utils
     from .version import __version__
