+++ nimare/annotate/text.py
@@ -9,7 +9,7 @@ import numpy as np
 import pandas as pd
 from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer
 
-from ..utils.utils import get_resource_path
+from ..utils import get_resource_path
 
 SPELL_DF = pd.read_csv(op.join(get_resource_path(), 'english_spellings.csv'),
                        index_col='UK')
