+++ nimare/annotate/topic/gclda.py
@@ -11,7 +11,7 @@ from scipy.stats import multivariate_normal
 
 from ...base import AnnotationModel
 from ...due import due, Doi
-from ...utils.utils import get_template
+from ...utils import get_template
 
 LGR = logging.getLogger(__name__)
