+++ nimare/io.py
@@ -13,12 +13,8 @@ import pandas as pd
 import numpy as np
 from pyneurovault import api
 
-from ..dataset import Database
+from .dataset import Database
 from .utils import get_resource_path, tal2mni
-
-
-__all__ = ['convert_sleuth', 'convert_sleuth_to_database',
-           'convert_sleuth_to_dict']
 
 
 def convert_sleuth_to_dict(text_file):
