--- nimare/utils/__init__.py
@@ -1,11 +1,0 @@
-"""
-Utility functions related to input/output, statistics, and general utilities.
-"""
-#from .io import convert_sleuth
-from .stats import one_way
-from .utils import get_template
-
-
-__all__ = [
-    'one_way', 'get_template',
-]
